requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
pydantic==2.10.6
firebase-admin==6.6.0
google-generativeai==0.8.4
python-dotenv==0.21.0
google-cloud-bigquery==3.27.0
sentence-transformers==4.0.2
scikit-learn==1.5.1
pdfplumber==0.11.5
pandas==2.2.2
python-multipart==0.0.19
tabulate==0.9.0
faiss-cpu==1.11.0
googletrans==3.1.0a0
langdetect==1.0.9
plotly==5.24.1
email_validator==2.2.0